app/requirements.txt
Flask==3.0.3
Flask_Cors==5.0.0
pymongo==4.10.1
python-dotenv==1.0.1
qrcode==7.4.2
gunicorn==20.1.0
Pillow
